fix_mistakes_with_tests

+++ Task_1_1_3/src/test/java/ru/nsu/laptev/ParserTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 class ParserTest {
     @Test
     void sampleTest() {
-        int e = Parser.parsing("x", "x = 10; x = 123; y = 0");
+        double e = Parser.parsing("x", "x = 10; x = 123; y = 0");
         assertEquals(123, e);
         try {
             e = Parser.parsing("y", "");
